REF make migrators in another job
