+++ .github/workflows/bot-make-migrators.yml
@@ -1,0 +1,78 @@
+name: bot-make-migrators
+
+on:
+  workflow_dispatch: null
+  schedule:
+    - cron: '*/15 * * * *'
+
+concurrency: bot-make-migrators
+
+jobs:
+  make-migrators:
+    name: make-migrators
+    runs-on: "ubuntu-latest"
+    defaults:
+      run:
+        shell: bash -leo pipefail {0}
+
+    steps:
+      - uses: actions/checkout@0ad4b8fadaa221de15dcec353f45205ec38ea70b # v4
+        with:
+          path: cf-scripts
+
+      - uses: mamba-org/setup-micromamba@422500192359a097648154e8db4e39bdb6c6eed7 # v1
+        with:
+          environment-file: cf-scripts/conda-lock.yml
+          environment-name: cf-scripts
+          condarc-file: cf-scripts/autotick-bot/condarc
+
+      - name: do local setup and stop me if needed
+        run: |
+          cd cf-scripts
+          python autotick-bot/stop_me_if_needed.py
+
+      - name: install bot code
+        if: success() && ! env.CI_SKIP
+        run: |
+          source cf-scripts/autotick-bot/install_bot_code.sh
+        env:
+          BOT_TOKEN: ${{ secrets.AUTOTICK_BOT_TOKEN }}
+
+      - name: make graph
+        if: success() && ! env.CI_SKIP
+        run: |
+          pushd cf-graph
+
+          conda-forge-tick make-migrators
+        env:
+          CF_TICK_GRAPH_DATA_BACKENDS: "${{ vars.CF_TICK_GRAPH_DATA_BACKENDS }}"
+          MONGODB_CONNECTION_STRING: ${{ secrets.MONGODB_CONNECTION_STRING }}
+
+      - name: deploy
+        if: github.ref == 'refs/heads/master' && always() && ! env.CI_SKIP
+        run: |
+          pushd cf-graph
+
+          export RUN_URL="https://github.com/regro/cf-scripts/actions/runs/${RUN_ID}"
+          conda-forge-tick deploy-to-github
+        env:
+          BOT_TOKEN: ${{ secrets.AUTOTICK_BOT_TOKEN }}
+          RUN_ID: ${{ github.run_id }}
+
+      - name: trigger next job
+        uses: benc-uk/workflow-dispatch@25b02cc069be46d637e8fe2f1e8484008e9e9609 # v1.2.3
+        if: github.ref == 'refs/heads/master' && ! cancelled() && ! failure() && ! env.CI_SKIP
+        with:
+          workflow: bot-make-migrators
+          ref: ${{ github.event.ref }}
+          token: ${{ secrets.AUTOTICK_BOT_TOKEN }}
+
+      - name: bump on fail
+        if: github.ref == 'refs/heads/master' && failure() && ! env.CI_SKIP
+        run: |
+          export ACTION_URL="https://github.com/regro/cf-scripts/actions/runs/${RUN_ID}"
+          python cf-scripts/autotick-bot/bump_bot_team.py
+        env:
+          BOT_TOKEN: ${{ secrets.AUTOTICK_BOT_TOKEN }}
+          RUN_ID: ${{ github.run_id }}
+          ACTION_NAME: ${{ github.workflow }}
